Consolidate library/items requests into single method
Adds a new method called get_abs_library_items which returns a list of all library items across all libraries.

This removes a good chunk of redundant code into a single reused method.

Removes the need for the user to specify a library ID and instead searches all of them.