Adds Collection/Playlist Sync (bi-directional)
Adds collections (which includes user playlists as a type of collection) as a sync-able column. From ABS to Calibre.

Updates the is_multiple attribute to be a tuple and support filtering the dropdown

Makes checkbox QLabels clickable such that they toggle the checkbox.

Updates ReadMe to reflect the new metadata