New and Improved Quick Link
Rewrote Quick Link from the ground up to make use of the audimeta API and match via the Audible ASIN for the Audiobookshelf item.

Basically, the new quick link is completely different from the old one but has the same goal, to try to make it easier to link your calibre books to Audiobookshelf.