Update QuickLink to use a worker and add ProgressDialog
QuickLink often resulted in the system being unresponsive for extended periods of time without any UX changes to show work was progressing. Moves the QuickLink loop to a worker in a background thread so it doesn't freeze calibre and adds a ProgressDialog to show the user progress in the linking.

Testing showed it takes about 0.55 seconds per book in QuickLink

Also updated some error messages and added time taken to the SyncCompletionDialog for QuickLink