Update sync_from_audiobookshelf to use a worker and add ProgressDialog
sync_from_audiobookshelf can result in the system being unresponsive for extended periods of time without any UX changes to show work was progressing. Moves the sync_from_audiobookshelf loop to a worker in a background thread so it doesn't freeze calibre and adds a ProgressDialog to show the user progress in the sync.

Basically the same as my prior commit for QuickLink. This one is less relevant imo but might as well. Hard to get a reference value here since this highly depends on number of values to update, it can vary quite a bit, but I did notice it hang sometimes with my library.

Also changed behavior of Narrator and Author to return None if not in ABS.