Speed Calculation Tweaks
Session handling is left to players which means there is some level of inconsistency in how sessions are are started/ended and what different attributes actually mean. In my case this meant that progression was the more accurate metric over time_listening as they were either the same or progression was more accurate with the exception of skipped content. I wish there was a more perfect solution but this is why all these numbers come with a big asterisk.

Also a fix for QL double click to open in ABS